+++ packages/taro-rn/src/utils/index.ts
@@ -85,59 +85,4 @@ export function isUrl (string: string): boolean {
   return false
 }
 
-interface CallbackManager {
-  add: (opt: any) => void;
-  remove: (opt: any) => void;
-  count: () => number;
-  trigger: (...args: any[]) => void;
-}
-
-export function createCallbackManager(): CallbackManager {
-  let callbacks: any[] = []
-
-  /**
-   * 添加回调
-   * @param {{ callback: function, ctx: any } | function} opt
-   */
-  const add = (opt) => {
-    callbacks.push(opt)
-  }
-
-  /**
-   * 移除回调
-   * @param {{ callback: function, ctx: any } | function} opt
-   */
-  const remove = (opt) => {
-    callbacks = callbacks.filter(callback => {
-      return callback !== opt
-    })
-  }
-
-  /**
-   * 获取回调函数数量
-   * @return {number}
-   */
-  const count = () => callbacks.length
-
-  /**
-   * 触发回调
-   * @param  {...any} args 回调的调用参数
-   */
-  const trigger = (...args) => {
-    callbacks.forEach(opt => {
-      if (typeof opt === 'function') {
-        opt(...args)
-      } else {
-        const { callback, ctx } = opt
-        callback.call(ctx, ...args)
-      }
-    })
-  }
-
-  return {
-    add,
-    remove,
-    count,
-    trigger
-  }
-}
+export * from './callbackManager'
